+++ daskms/table_proxy.py
@@ -235,6 +235,24 @@ def _iswriteable(table_future):
     return table_future.result().iswritable()
 
 
+def _table_close(table_future):
+    table_future.result().close()
+
+
+# This finalizer submits the table closing operation
+# on the Executor associated with the TableProxy
+# Note that this operation **must** happen before
+# shutdown is called on the Executor's internal
+# ThreadPoolExecutor object or deadlock can occur
+# See https://codewithoutrules.com/2017/08/16/concurrency-python/
+# for further insight.
+# The fact that the Executor is an argument to weakref.finalize implies
+# that the Executor's finalizer, and by further implication,
+# the ThreadPoolExecutor.shutdown has not yet been called.
+def _finaliser(ex, table_future):
+    ex.impl.submit(_table_close, table_future)
+
+
 class TableProxy(object, metaclass=TableProxyMetaClass):
     """
     Proxies calls to a :class:`pyrap.tables.table` object via
@@ -274,13 +292,14 @@ class TableProxy(object, metaclass=TableProxyMetaClass):
         kwargs = kwargs.copy()
         self._ex_key = kwargs.pop("__executor_key__", STANDARD_EXECUTOR)
 
-        ex = Executor(key=self._ex_key)
-        self._table_future = table = ex.impl.submit(factory, *args, **kwargs)
-
         # Store a reference to the Executor wrapper class
         # so that the Executor is retained while this TableProxy
         # still lives
-        self._ex_wrapper = ex
+        self._ex_wrapper = ex = Executor(key=self._ex_key)
+        self._table_future = table = ex.impl.submit(factory, *args, **kwargs)
+
+        weakref.finalize(self, _finaliser, ex, table)
+
         # Reference to the internal ThreadPoolExecutor
         self._ex = ex.impl